Refactor API.ts god object into separate grouped files. Added icons to the UI.
