+++ frontend/src/pages/ClassHome.tsx
@@ -4,7 +4,8 @@ import Button from "../components/Button";
 import "./ClassHome.css";
 import { useNavigate, useParams } from "react-router-dom";
 import { useState, useEffect } from "react";
-import { listAssignments, listClasses, deleteAssignment, enrollStudentsByEmail } from "../util/api";
+import { listAssignments, deleteAssignment } from "../util/api_client/assignments";
+import { listClasses, enrollStudentsByEmail } from "../util/api_client/classes";
 import TabNavigation from "../components/TabNavigation";
 import { importCSV } from "../util/csv";
 import StatusMessage from "../components/StatusMessage";
