+++ frontend/src/pages/LoginPage.css
@@ -5,7 +5,6 @@
   justify-content: center;
 
   height: 100vh;
-  width: 100vw;
   overflow: hidden;
 
   /* Negative margin to offset the "inner" padding and ensure the background covers the entire viewport */
